Added Collision Detection (BUGGY)
Added collision detection to the player so that they don't fall through the floor. IT'S BUGGY AND I MAY FIX IT LATER, since the player gets stuck on the floor when jumping too high, and they would need to jup twice to get out of the floor and jump in the air.