Added Collision Detection (BUGGY)
Fixed a bit the collision detection. The player still visibly looks like his feet are stuck on the ground, and every 20 jumps or so, the player will have to jump twice to escape from the floor. However, the game's still playable, and the bug may take too much time to fix, so I'll leave it as is.