Fixed platform bug
I fixed the bug where the player was being spawned on top of the platforms if walking below them, and I fixed the bug where the player would spawn on top of the medium and tall platform if the jumped below them. I also splitted the platform sprite into a top half and a bottom half.