Added ability to open chests
The player is able to open chests. If the user presses the E key at any moment, the chest will change from the default closed chest sprite to the opened chest sprite. However, the opened chest sprite is taller than the closed one, so it looks like the chest sinks through the floor. Also, I only want to open the chest if I'm near it.