Fixed Chest Bug (BUGGY CODE)
I fixed a bug that caused all of the chests to be opened at the same time if I only opened 1 chest. However, the treasures (like the bronze ring) are not being rendered anymore. Also, before fixing this bug, I noticed that, even if a chest was opened, if I kept pressing the E key on that opened chest, the treasure would reappear for 3 seconds. I need to fix those 2 treasure bugs.