Fixing Bug From treasure.lua (BUGGY)
I fixed the bug that caused all of the treasures to pop up on the UI immediately after entering a new room after opening a chest for all chests. However, I have new bug that makes all of the treasure icons on the UI whenever I go to the main hub.