Added Cage on Lever.lua
The game will now render a cage on top of a chest on room 2. If a lever is activated, that cage will disappear. I need to now make the chest inside of the cage to be locked and impossible to open. The only way to open it will be by activating the lever and making the chest disappear.