Fixed lever bug (BUGGY)
I fixed a bug in which a the lever from another room was present but invisible in the other rooms (for instance, that the lever from Room 3 was in room 2 but invisible). This caused that, if I stood in the position of the invisible lever and pressed E, that invisible lever would be activated and the treasure from the other room would be unlocked. However, I still need to lock the chest from room 3 so that it won't be unlocked unless I pull room 3's lever.