Updated Stalactites.lua (BUG FIX)
I fixed a bug in which, since the stalactites will always be present in all rooms, even if they aren't supposed to be there, the player will be able to jump on top of them while they are invisible, even on rooms that shouldn't have stalactites. I just hid the stalactites below the viewport's height to avoid any problems.