Bug Fix in Treasure.lua
I fixed the bug that caused the treasure of a new room to be rendered even if the chest was never opened if the player opened a chest and they immediately afterwards entered a new treasure room. However, this is only if the player travels to the hub first BEFORE entering the new treasure room. If I skip from one treasure room to the other using chests (like pressing the 2 key after opening the chest from any other room other than room 2), the bug will still be there.