Added In-game Victory Message (Buggy)
I'm now rendering the victory message into the game. However, the text is white and too hard to read against the room's backgrounds. Also, the text is not appearing in the center of the screen. So, I may add rectangle that will work as a background to make the message essy to read.